+++ packages/core-js/stage/0.js
@@ -4,5 +4,7 @@ require('../modules/esnext.math.isubh');
 require('../modules/esnext.math.imulh');
 require('../modules/esnext.math.umulh');
 require('../modules/esnext.string.at');
+require('../modules/esnext.symbol.pattern-match');
+require('../modules/esnext.symbol.pattern-value');
 
 module.exports = require('./1');